# docs/contracts/negociacao.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Negociação Service",
    "description": "Microsserviço de Negociação/Leilão do Portal B2B. Gerencia processo_negociacao e lance, publica negociacao_fechada.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/processos": {
      "get": {
        "tags": [
          "negociacao"
        ],
        "summary": "List Processos",
        "operationId": "list_processos_processos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "modo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modo"
            }
          },
          {
            "name": "produto_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Produto Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessoOut"
                  },
                  "title": "Response List Processos Processos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/processos/{processo_id}": {
      "get": {
        "tags": [
          "negociacao"
        ],
        "summary": "Get Processo",
        "operationId": "get_processo_processos__processo_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "processo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Processo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessoDetalhadoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/processos/{processo_id}/lances": {
      "post": {
        "tags": [
          "negociacao"
        ],
        "summary": "Registrar Lance",
        "operationId": "registrar_lance_processos__processo_id__lances_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "processo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Processo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/processos/{processo_id}/fechar": {
      "post": {
        "tags": [
          "negociacao"
        ],
        "summary": "Fechar Processo",
        "operationId": "fechar_processo_processos__processo_id__fechar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "processo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Processo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LanceInput": {
        "properties": {
          "valor_unitario": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Valor Unitario"
          },
          "quantidade": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantidade"
          },
          "empresa_id_override": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa Id Override"
          }
        },
        "type": "object",
        "required": [
          "valor_unitario",
          "quantidade"
        ],
        "title": "LanceInput"
      },
      "LanceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "processo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Processo Id"
          },
          "empresa_id": {
            "type": "string",
            "format": "uuid",
            "title": "Empresa Id"
          },
          "valor_unitario": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Valor Unitario"
          },
          "quantidade": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantidade"
          },
          "data_lance": {
            "type": "string",
            "format": "date-time",
            "title": "Data Lance"
          }
        },
        "type": "object",
        "required": [
          "id",
          "processo_id",
          "empresa_id",
          "valor_unitario",
          "quantidade",
          "data_lance"
        ],
        "title": "LanceOut"
      },
      "ProcessoDetalhadoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "produto_id": {
            "type": "string",
            "format": "uuid",
            "title": "Produto Id"
          },
          "modo": {
            "type": "string",
            "enum": [
              "direto",
              "leilao_direto",
              "leilao_reverso"
            ],
            "title": "Modo"
          },
          "status": {
            "type": "string",
            "enum": [
              "ABERTO",
              "FECHADA",
              "CANCELADO"
            ],
            "title": "Status"
          },
          "data_inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Data Inicio"
          },
          "data_fim": {
            "type": "string",
            "format": "date-time",
            "title": "Data Fim"
          },
          "valor_reserva": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Reserva"
          },
          "lances": {
            "items": {
              "$ref": "#/components/schemas/LanceOut"
            },
            "type": "array",
            "title": "Lances"
          }
        },
        "type": "object",
        "required": [
          "id",
          "produto_id",
          "modo",
          "status",
          "data_inicio",
          "data_fim",
          "valor_reserva"
        ],
        "title": "ProcessoDetalhadoOut"
      },
      "ProcessoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "produto_id": {
            "type": "string",
            "format": "uuid",
            "title": "Produto Id"
          },
          "modo": {
            "type": "string",
            "enum": [
              "direto",
              "leilao_direto",
              "leilao_reverso"
            ],
            "title": "Modo"
          },
          "status": {
            "type": "string",
            "enum": [
              "ABERTO",
              "FECHADA",
              "CANCELADO"
            ],
            "title": "Status"
          },
          "data_inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Data Inicio"
          },
          "data_fim": {
            "type": "string",
            "format": "date-time",
            "title": "Data Fim"
          },
          "valor_reserva": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Reserva"
          }
        },
        "type": "object",
        "required": [
          "id",
          "produto_id",
          "modo",
          "status",
          "data_inicio",
          "data_fim",
          "valor_reserva"
        ],
        "title": "ProcessoOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}